% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_clip.R
\name{sum_clip}
\alias{sum_clip}
\title{sum_clip}
\usage{
sum_clip(big.mark = TRUE, return_x = FALSE)
}
\arguments{
\item{big.mark}{bool. If TRUE commas are inserted to differentiate long digits}

\item{return_x}{bool. If you want to see the input vector}
}
\description{
Sum values in a clipboard taken from excel'
}
